% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackopt-package.R
\docType{package}
\name{trackopt-package}
\alias{trackopt}
\alias{trackopt-package}
\title{trackopt: Track Numerical Optimization}
\description{
Tracks parameter values, gradients, and Hessians at each iteration of numerical optimizers. Useful for analyzing optimization progress, diagnosing issues, and studying convergence behavior.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/trackopt}
  \item Report bugs at \url{https://github.com/loelschlaeger/trackopt/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com}

}
\keyword{internal}
